Convert StartupAnalyzerTest to the new test framework
