--- src/Analyzers/Analyzers/test/TestFiles/StartupAnalyzerTest/MvcOptions_UseMvcWithDefaultRouteAndAddMvcOptionsEndpointRoutingDisabled.cs
@@ -1,21 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Microsoft.AspNetCore.Analyzers.TestFiles.StartupAnalyzerTest
-{
-    public class MvcOptions_UseMvcWithDefaultRouteAndAddMvcOptionsEndpointRoutingDisabled
-    {
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMvc().AddMvcOptions(options => options.EnableEndpointRouting = false);
-        }
-
-        public void Configure(IApplicationBuilder app)
-        {
-            app.UseMvcWithDefaultRoute();
-        }
-    }
-}
